lumin_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pyjwt
cryptography
apscheduler
firebase-admin